% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_move_types.R
\name{assign_movement_types}
\alias{assign_movement_types}
\title{Assign movement types.}
\usage{
assign_movement_types(df)
}
\arguments{
\item{df}{A dataframe with generation, id, and movement weights.}
}
\value{
Nothing. Assigns strategy by reference. See data.table.
}
\description{
Assign movement types.
}
