% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_social_strat.R
\name{get_disease_burden}
\alias{get_disease_burden}
\title{Get disease burden per social class.}
\usage{
get_disease_burden(df)
}
\arguments{
\item{df}{A dataframe with social weights and ecological outcomes.}
}
\value{
A dataframe of disease burden, the sum of the time infected per
strategy in each generation.
}
\description{
Get disease burden per social class.
}
