% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_network_metrics.R
\name{get_edge_distance}
\alias{get_edge_distance}
\title{Get landscape distance between individual positions.}
\usage{
get_edge_distance(g)
}
\arguments{
\item{g}{An igraph or tidygraph object.}
}
\value{
A data.table of edges with landscapes distances between nodes.
}
\description{
Get landscape distance between individual positions.
}
