% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_network_metrics.R
\name{get_global_efficiency}
\alias{get_global_efficiency}
\title{Get network global efficiency.}
\usage{
get_global_efficiency(g, vertices = 500)
}
\arguments{
\item{g}{An igraph or tidygraph object.}

\item{vertices}{Number of nodes, 500 by default.}
}
\value{
A numeric global efficiency value.
}
\description{
Get network global efficiency.
}
