% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_social_strat.R
\name{get_social_outcomes}
\alias{get_social_outcomes}
\title{Get ecological outcomes of evolved social strategies.}
\usage{
get_social_outcomes(df)
}
\arguments{
\item{df}{A dataframe with social weights and ecological outcomes.}
}
\value{
A dataframe with outcomes summarised by generation and social
strategy.
}
\description{
Get ecological outcomes of evolved social strategies.
}
