% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_network_metrics.R
\name{get_strategy_networks}
\alias{get_strategy_networks}
\title{Get connections between social strategies.}
\usage{
get_strategy_networks(g)
}
\arguments{
\item{g}{An igraph or tidygraph object, with node data containing social
strategy information.}
}
\value{
A tidygraph network of connections between social strategies.
}
\description{
Get connections between social strategies.
}
