% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_disease_cost.R
\name{get_total_disease_cost}
\alias{get_total_disease_cost}
\title{Get disease cost from time infected.}
\usage{
get_total_disease_cost(df, cost = 0.25)
}
\arguments{
\item{df}{A dataframe with time infected.}

\item{cost}{Disease cost per timestep.}
}
\value{
Assign total disease cost per individual.
}
\description{
Get disease cost from time infected.
}
