% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_disease_cost.R
\name{get_total_energy}
\alias{get_total_energy}
\title{Get agent energy.}
\usage{
get_total_energy(df, cost_disease = 0.25)
}
\arguments{
\item{df}{A dataframe with agent intake and time infected.}

\item{cost_disease}{The cost of the disease.}
}
\value{

}
\description{
Get agent energy.
}
