% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_network_metrics.R
\name{get_wrapped_distance}
\alias{get_wrapped_distance}
\title{Get wrapped distance.}
\usage{
get_wrapped_distance(x1, y1, x2, y2, max_dist = 50)
}
\arguments{
\item{x1}{X coord 1.}

\item{y1}{Y coord 1.}

\item{x2}{X coord 2.}

\item{y2}{Y coord 2.
#' @param vertices Number of nodes, 500 by default.}
}
\value{
A numeric global efficiency value.
}
\description{
Get wrapped distance.
}
